Rename function and class containing `from_json`
