+++ src/main/cpp/src/JSONUtilsJni.cpp
@@ -15,6 +15,7 @@
  */
 
 #include "cudf_jni_apis.hpp"
+#include "from_json.hpp"
 #include "get_json_object.hpp"
 
 #include <cudf/strings/strings_column_view.hpp>
@@ -136,6 +137,19 @@ Java_com_nvidia_spark_rapids_jni_JSONUtils_getJsonObjectMultiplePaths(JNIEnv* en
       return cudf::jni::release_as_jlong(col);
     });
     return out_handles.get_jArray();
+  }
+  CATCH_STD(env, 0);
+}
+
+JNIEXPORT jlong JNICALL Java_com_nvidia_spark_rapids_jni_JSONUtils_extractRawMapFromJsonString(
+  JNIEnv* env, jclass, jlong input_handle)
+{
+  JNI_NULL_CHECK(env, input_handle, "json_column_handle is null", 0);
+
+  try {
+    cudf::jni::auto_set_device(env);
+    auto const input = reinterpret_cast<cudf::column_view const*>(input_handle);
+    return cudf::jni::ptr_as_jlong(spark_rapids_jni::from_json_to_raw_map(*input).release());
   }
   CATCH_STD(env, 0);
 }
