+++ src/main/cpp/src/from_json.hpp
@@ -14,23 +14,21 @@
  * limitations under the License.
  */
 
-#include "map_utils.hpp"
+#pragma once
 
-#include <cudf_jni_apis.hpp>
-#include <dtype_utils.hpp>
+#include <cudf/column/column_view.hpp>
+#include <cudf/utilities/default_stream.hpp>
 
-extern "C" {
+#include <rmm/cuda_stream_view.hpp>
+#include <rmm/resource_ref.hpp>
 
-JNIEXPORT jlong JNICALL Java_com_nvidia_spark_rapids_jni_MapUtils_extractRawMapFromJsonString(
-  JNIEnv* env, jclass, jlong input_handle)
-{
-  JNI_NULL_CHECK(env, input_handle, "json_column_handle is null", 0);
+#include <memory>
 
-  try {
-    cudf::jni::auto_set_device(env);
-    auto const input = reinterpret_cast<cudf::column_view const*>(input_handle);
-    return cudf::jni::ptr_as_jlong(spark_rapids_jni::from_json(*input).release());
-  }
-  CATCH_STD(env, 0);
-}
-}
+namespace spark_rapids_jni {
+
+std::unique_ptr<cudf::column> from_json_to_raw_map(
+  cudf::column_view const& input,
+  rmm::cuda_stream_view stream      = cudf::get_default_stream(),
+  rmm::device_async_resource_ref mr = rmm::mr::get_current_device_resource());
+
+}  // namespace spark_rapids_jni
