+++ src/main/java/com/nvidia/spark/rapids/jni/JSONUtils.java
@@ -40,16 +40,16 @@ public class JSONUtils {
     private final int index;
 
     public PathInstructionJni(PathInstructionType type, String name, long index) {
-      this.type = (byte)type.ordinal();
+      this.type = (byte) type.ordinal();
       this.name = name;
       if (index > Integer.MAX_VALUE) {
         throw new IllegalArgumentException("index is too large " + index);
       }
-      this.index = (int)index;
+      this.index = (int) index;
     }
 
     public PathInstructionJni(PathInstructionType type, String name, int index) {
-      this.type = (byte)type.ordinal();
+      this.type = (byte) type.ordinal();
       this.name = name;
       this.index = index;
     }
@@ -137,6 +137,30 @@ public class JSONUtils {
     return ret;
   }
 
+
+  /**
+   * Extract key-value pairs for each output map from the given json strings. These key-value are
+   * copied directly as substrings of the input without any type conversion.
+   * <p>
+   * Since there is not any validity check, the output of this function may be different from
+   * what generated by Spark's `from_json` function. Situations that can lead to
+   * different/incorrect outputs may include:<br>
+   * - The value in the input json string is invalid, such as 'abc' value for an integer key.<br>
+   * - The value string can be non-clean format for floating-point type, such as '1.00000'.
+   * <p>
+   * The output of these situations should all be NULL or a value '1.0', respectively. However, this
+   * function will just simply copy the input value strings to the output.
+   *
+   * @param jsonColumn The input strings column in which each row specifies a json object.
+   * @return A map column (i.e., a column of type {@code List<Struct<String,String>>}) in
+   * which the key-value pairs are extracted directly from the input json strings.
+   */
+  public static ColumnVector extractRawMapFromJsonString(ColumnView jsonColumn) {
+    assert jsonColumn.getType().equals(DType.STRING) : "Input type must be String";
+    return new ColumnVector(extractRawMapFromJsonString(jsonColumn.getNativeView()));
+  }
+
+
   private static native int getMaxJSONPathDepth();
 
   private static native long getJsonObject(long input,
@@ -151,4 +175,7 @@ public class JSONUtils {
                                                           int[] pathOffsets,
                                                           long memoryBudgetBytes,
                                                           int parallelOverride);
+
+
+  private static native long extractRawMapFromJsonString(long jsonColumnHandle);
 }
